Ship the new GUI!
- Integrate the Textual app into the framework
- Add fancy details to the UI
- Improve the API of the GUI
- Refactor a little
