+++ utils/gui.py
@@ -1,182 +1,183 @@
-#! /usr/bin/env python3
-# vim:fenc=utf-8
-#
-# Copyright © 2024 Théo Morales <theo.morales.fr@gmail.com>
-#
-# Distributed under terms of the MIT license.
+# Let's use Textual to rewrite the GUI with better features.
 
-"""
-The fancy new GUI.
-"""
-
-import random
-from collections import abc, namedtuple
+import asyncio
+from collections import abc
 from datetime import datetime
+from enum import Enum
 from functools import partial
+from itertools import cycle
+from random import random
 from time import sleep
 from typing import (
+    Any,
     Callable,
     Iterable,
     Iterator,
-    List,
     Optional,
     Sequence,
     Tuple,
-    TypeVar,
 )
 
-import plotext as plt
-import rich
+import numpy as np
 import torch
-from rich import box
-from rich.align import Align
-from rich.ansi import AnsiDecoder
-from rich.console import Group
-from rich.jupyter import JupyterMixin
-from rich.layout import Layout
-from rich.live import Live
-from rich.panel import Panel
-from rich.progress import (
-    BarColumn,
-    Progress,
-    SpinnerColumn,
-    TaskID,
-    TaskProgressColumn,
-    TextColumn,
-    TimeRemainingColumn,
-)
-from rich.style import Style
-from rich.table import Table
+import torch.multiprocessing as mp
+from rich.console import Group, RenderableType
+from rich.pretty import Pretty
 from rich.text import Text
+from textual.app import App, ComposeResult
+from textual.containers import Center
+from textual.reactive import var
+from textual.widgets import (
+    Footer,
+    Header,
+    Label,
+    Placeholder,
+    ProgressBar,
+    RichLog,
+    Static,
+)
+from textual_plotext import PlotextPlot
 from torch.utils.data.dataloader import DataLoader
 from torchvision.datasets import MNIST
 from torchvision.transforms.functional import to_tensor
 
-if __name__ != "__main__":
-    from conf import project as project_conf
-    from utils.helpers import BestNModelSaver
-else:
-    project_conf = namedtuple("project_conf", "Theme")(
-        Theme=namedtuple("Theme", "TRAINING VALIDATION TESTING")(
-            TRAINING=namedtuple("value", "value")("blue"),
-            VALIDATION=namedtuple("value", "value")("green"),
-            TESTING=namedtuple("value", "value")("cyan"),
-        )
-    )
-    BestNModelSaver = TypeVar("BestNModelSaver")
 
+class PlotterWidget(PlotextPlot):
+    marker: var[str] = var("sd")
 
-class PlotextMixin(JupyterMixin):
-    def __init__(self, p_make_plot):
-        self.decoder = AnsiDecoder()
-        self.mk_plot = p_make_plot
+    """The type of marker to use for the plot."""
 
-    def __rich_console__(self, console, options):
-        self.width = options.max_width or console.width
-        self.height = options.height or console.height
-        canvas = self.mk_plot(width=self.width, height=self.height)
-        self.rich_canvas = Group(*self.decoder.decode(canvas))
-        yield self.rich_canvas
+    def __init__(
+        self,
+        title: str,
+        use_log_scale: bool = False,
+        *,
+        name: str | None = None,
+        id: str | None = None,  # pylint:disable=redefined-builtin
+        classes: str | None = None,
+        disabled: bool = False,
+    ) -> None:
+        """Initialise the training curves plotter widget.
 
+        Args:
+            name: The name of the plotter widget.
+            id: The ID of the plotter widget in the DOM.
+            classes: The CSS classes of the plotter widget.
+            disabled: Whether the plotter widget is disabled or not.
+        """
+        super().__init__(name=name, id=id, classes=classes, disabled=disabled)
+        self._title = title
+        self._log_scale = use_log_scale
+        self._train_losses: list[float] = []
+        self._val_losses: list[float] = []
+        self._start_epoch = 0
+        self._epoch = 0
 
-# TODO: Make it a singleton so we can print from anywhere in the code, without passing a reference
-# around.
-class GUI:
-    def __init__(self, run_name: str, plot_log_scale: bool) -> None:
-        self._run_name = run_name
-        self._plot_log_scale = plot_log_scale
-        self._starting_epoch = 0  # TODO:
-        self._layout = Layout()
-        self._layout.split(
-            Layout(name="header", size=2),
-            Layout(name="main", ratio=1),
-            Layout(name="footer", size=2),
-        )
-        self._layout["main"].split_row(
-            Layout(name="body", ratio=3, minimum_size=60),
-            Layout(name="side"),
-        )
-        self._plot = Panel(
-            Align.center(
-                Text(
-                    "Waiting for training curves...",
-                    justify="center",
-                    style=Style(color="blue", bold=True),
-                ),
-                vertical="middle",
-            ),
-            title="Training curves",
-            expand=True,
-        )
-        self._layout["body"].update(self._plot)
-        self._live = Live(self._layout, screen=True)
-        self._console = self._live.console
-        self._pbar = Progress(
-            SpinnerColumn(spinner_name="monkey"),
-            TextColumn(
-                "[progress.description]{task.description} \[loss={task.fields[loss]:.3f}]"
-            ),
-            BarColumn(),
-            TaskProgressColumn(),
-            TimeRemainingColumn(),
-            # expand=True,
-        )
-        self._main_progress = Panel(  # TODO:
-            self._pbar,
-            title="Training epoch ?/?",
-            expand=True,
-        )
-        self._layout["footer"].update(self._pbar)
-        run_color = f"color({hash(run_name) % 255})"
-        background_color = f"color({(hash(run_name) + 128) % 255})"
-        self._layout["header"].update(
-            Text(
-                f"Running {run_name}",
-                style=f"bold {run_color} on {background_color}",
-                justify="center",
+    def on_mount(self) -> None:
+        """Plot the data using Plotext."""
+        self.plt.title(self._title)
+        self.plt.xlabel("Epoch")
+        if self._log_scale:
+            self.plt.ylabel("Loss (log scale)")
+            self.plt.yscale("log")
+        else:
+            self.plt.ylabel("Loss")
+        self.plt.grid(True, True)
+
+    def replot(self) -> None:
+        """Redraw the plot."""
+        self.plt.clear_data()
+        if self._log_scale and (
+            self._train_losses[-1] <= 0 or self._val_losses[-1] <= 0
+        ):
+            raise ValueError(
+                "Cannot plot on a log scale if there are non-positive losses."
             )
-        )
-        self._logger = Table.grid(padding=0)
-        self._logger.add_column(no_wrap=False)
-        self._layout["side"].update(
-            Panel(
-                self._logger, title="Logs", border_style="bright_red", box=box.ROUNDED
+        if len(self._train_losses) > 0:
+            assert len(self._val_losses) == len(self._train_losses)
+            self.plt.plot(
+                list(range(self._start_epoch, self._epoch + 1)),
+                self._train_losses,
+                color="blue",  # TODO: Theme
+                label="Training loss",
+                marker=self.marker,
             )
+            self.plt.plot(
+                list(range(self._start_epoch, self._epoch + 1)),
+                self._val_losses,
+                color="green",  # TODO: Theme
+                label="Validation loss",
+                marker=self.marker,
+            )
+        self.refresh()
+
+    def set_start_epoch(self, start_epoch: int):
+        self._start_epoch = start_epoch
+
+    def update(
+        self, epoch: int, train_loss: float, val_loss: Optional[float] = None
+    ) -> None:
+        """Update the data for the training curves plot.
+
+        Args:
+            epoch: (int) The current epoch number.
+            train_loss: (float) The last training loss.
+            val_loss: (float) The last validation loss.
+        """
+        self._epoch = epoch
+        self._train_losses.append(train_loss)
+        self._val_losses.append(
+            val_loss if val_loss is not None else self._val_losses[-1]
         )
-        self.tasks = {
-            "training": self._pbar.add_task(
-                f"[{project_conf.Theme.TRAINING.value}]Training",
-                visible=False,
-                loss=torch.inf,
-            ),
-            "validation": self._pbar.add_task(
-                f"[{project_conf.Theme.VALIDATION.value}]Validation",
-                visible=False,
-                loss=torch.inf,
-            ),
-            "testing": self._pbar.add_task(
-                f"[{project_conf.Theme.TESTING.value}]Testing",
-                visible=False,
-                loss=torch.inf,
-            ),
-        }
+        self.replot()
 
-    @property
-    def console(self):
-        return self._console
+    def _watch_marker(self) -> None:
+        """React to the marker being changed."""
+        self.replot()
 
-    def open(self) -> None:
-        self._live.__enter__()
 
-    def close(self) -> None:
-        self._live.__exit__(None, None, None)
+# TODO: Also make a Rich renderable for Tensors (using tables?)
 
-    def _track_iterable(self, iterable, task, total) -> Tuple[Iterable, Callable]:
+
+class Task(Enum):
+    IDLE = -1
+    TRAINING = 0
+    VALIDATION = 1
+    TESTING = 2
+
+
+class DatasetProgressBar(Static):
+    """A progress bar for PyTorch dataloader iteration."""
+
+    DESCRIPTIONS = {
+        Task.IDLE: Text("Waiting for work..."),
+        Task.TRAINING: Text("Training: ", style="bold blue"),
+        Task.VALIDATION: Text("Validation: ", style="bold green"),
+        Task.TESTING: Text("Testing: ", style="bold yellow"),
+    }
+
+    def compose(self) -> ComposeResult:
+        # with Horizontal():
+        with Center():
+            yield Label(self.DESCRIPTIONS[Task.IDLE], id="progress_label")
+            yield ProgressBar()
+
+    def track_iterable(
+        self,
+        iterable: Iterable | Sequence | Iterator | DataLoader,
+        task: Task,
+        total: int,
+    ) -> Tuple[Iterable, Callable]:
         class LossHook:
             def __init__(self):
                 self._loss = None
 
-            def update_loss_hook(self, loss: float):
+            def update_loss_hook(
+                self, loss: float, min_val_loss: Optional[float] = None
+            ) -> None:
+                """Update the loss value in the progress bar."""
+                # TODO: min_val_loss during validation, val_loss during training. Ideally the
+                # second parameter would be super flexible (use a dict then).
                 self._loss = loss
 
         class SeqWrapper(abc.Iterator, LossHook):
@@ -226,22 +227,22 @@ class GUI:
                     self._reset_hook()
                     raise StopIteration
 
-        def update_hook(task_id: TaskID, loss: Optional[float] = None):
-            self._pbar.advance(task_id)
+        def update_hook(loss: Optional[float] = None):
+            self.query_one(ProgressBar).advance()
             if loss is not None:
-                self._pbar.tasks[task_id].fields["loss"] = loss
-            # TODO: Nice progress panel with overall progress and epoch progress
-            # self._main_progress = Panel(self._pbar, title="Training epoch ?/?")
-            # self._layout["footer"].update(self._main_progress)
-            # self._live.refresh()
+                plabel: Label = self.query_one("#progress_label")  # type: ignore
+                plabel.update(self.DESCRIPTIONS[task] + f"[loss={loss:.4f}]")
 
-        def reset_hook(task_id: TaskID, total: int):
-            self._pbar.reset(task_id, total=total, visible=False)
+        def reset_hook(total: int):
+            sleep(0.5)
+            self.query_one(ProgressBar).update(total=100, progress=0)
+            plabel: Label = self.query_one("#progress_label")  # type: ignore
+            plabel.update(self.DESCRIPTIONS[Task.IDLE])
 
         wrapper = None
         update_p, reset_p = (
-            partial(update_hook, task_id=task),
-            partial(reset_hook, task, total),
+            partial(update_hook),
+            partial(reset_hook, total),
         )
         if isinstance(iterable, abc.Sequence):
             wrapper = SeqWrapper(
@@ -261,167 +262,199 @@ class GUI:
             raise ValueError(
                 f"iterable must be a Sequence or an Iterator, got {type(iterable)}"
             )
-        self._pbar.reset(task, total=total, visible=True)
+        self.query_one(ProgressBar).update(total=total, progress=0)
+        plabel: Label = self.query_one("#progress_label")  # type: ignore
+        plabel.update(self.DESCRIPTIONS[task])
         return wrapper, wrapper.update_loss_hook
 
+
+class GUI(App):
+    """A Textual app to serve as *useful* GUI/TUI for my pytorch-based micro framework."""
+
+    CSS_PATH = "style.css"
+
+    BINDINGS = [
+        ("q", "quit", "Quit"),
+        ("d", "toggle_dark", "Toggle dark mode"),
+        ("p", "marker", "Change plotter style"),
+        ("ctrl+z", "suspend_progress"),
+    ]
+
+    MARKERS = {
+        "dot": "Dot",
+        "hd": "High Definition",
+        "fhd": "Higher Definition",
+        "braille": "Braille",
+        "sd": "Standard Definition",
+    }
+
+    marker: var[str] = var("hd")
+
+    def __init__(self, run_name: str, log_scale: bool) -> None:
+        """Initialise the application."""
+        super().__init__()
+        self._markers = cycle(self.MARKERS.keys())
+        self._log_scale = log_scale
+        self.run_name = run_name
+
+    def compose(self) -> ComposeResult:
+        yield Header()
+        yield PlotterWidget(
+            title=f"Trainign curves for {self.run_name}",
+            use_log_scale=self._log_scale,
+            classes="box",
+        )
+        yield RichLog(
+            highlight=True, markup=True, wrap=True, id="logger", classes="box"
+        )
+        yield DatasetProgressBar()
+        yield Placeholder(classes="box")
+        yield Footer()
+
+    def on_mount(self):
+        self.query_one(PlotterWidget).loading = True
+
+    def action_toggle_dark(self) -> None:
+        self.dark = not self.dark
+
+    def watch_marker(self) -> None:
+        """React to the marker type being changed."""
+        self.sub_title = self.MARKERS[self.marker]
+        self.query_one(PlotterWidget).marker = self.marker
+
+    def action_marker(self) -> None:
+        """Cycle to the next marker type."""
+        self.marker = next(self._markers)
+
+    def print(self, message: Any):
+        logger: RichLog = self.query_one(RichLog)
+        if isinstance(message, (RenderableType, str)):
+            logger.write(
+                Group(
+                    Text(
+                        datetime.now().strftime("[%H:%M] "),
+                        style="dim cyan",
+                        end="",
+                    ),
+                    message,
+                ),
+            )
+        else:
+            ppable, pp_msg = True, None
+            try:
+                pp_msg = Pretty(message)
+            except Exception:
+                ppable = False
+            if ppable and pp_msg is not None:
+                logger.write(
+                    Group(
+                        Text(
+                            datetime.now().strftime("[%H:%M] "),
+                            style="dim cyan",
+                            end="",
+                        ),
+                        Text(str(type(message)) + " ", style="italic blue", end=""),
+                        pp_msg,
+                    )
+                )
+            else:
+                try:
+                    logger.write(
+                        Group(
+                            Text(
+                                datetime.now().strftime("[%H:%M] "),
+                                style="dim cyan",
+                                end="",
+                            ),
+                            message,
+                        ),
+                    )
+                except Exception as e:
+                    logger.write(
+                        Group(
+                            Text(
+                                datetime.now().strftime("[%H:%M] "),
+                                style="dim cyan",
+                                end="",
+                            ),
+                            Text("Logging error: ", style="bold red"),
+                            Text(str(e), style="bold red"),
+                        )
+                    )
+
     def track_training(self, iterable, total: int) -> Tuple[Iterable, Callable]:
-        task = self.tasks["training"]
-        return self._track_iterable(iterable, task, total)
+        """Return an iterable that tracks the progress of the training process, and a progress bar
+        hook to update the loss value at each iteration."""
+        return self.query_one(DatasetProgressBar).track_iterable(
+            iterable, Task.TRAINING, total
+        )
 
     def track_validation(self, iterable, total: int) -> Tuple[Iterable, Callable]:
-        task = self.tasks["validation"]
-        return self._track_iterable(iterable, task, total)
+        """Return an iterable that tracks the progress of the validation process, and a progress bar
+        hook to update the loss value at each iteration."""
+        return self.query_one(DatasetProgressBar).track_iterable(
+            iterable, Task.VALIDATION, total
+        )
 
     def track_testing(self, iterable, total: int) -> Tuple[Iterable, Callable]:
-        task = self.tasks["testing"]
-        return self._track_iterable(iterable, task, total)
-
-    def print_header(self, text: str):
-        self._layout["header"].update(text)
-
-    def print(self, text: str | Text):
-        """
-        Print text to the side panel.
-        """
-        # TODO: Use a fifo and pop the first item if we can't view the enw item.
-        try:
-            self._logger.add_row(
-                Text(datetime.now().strftime("[%H:%M] "), style="dim cyan"), text
-            )
-            # TODO: Compute the max number of displayable rows. We have access to the height:
-            # height = self._console.size.height
-            # But what about the width of the column? Can we get that anywhere?
-            # Then, we need to compute the width of each row to see if it overflows and by how many
-            # lines.
-            # Finally, we'll have the percentage of height that our rows take, and if it's above
-            # 100%-threshold, we remove the first row.
-        except rich.errors.NotRenderableError as e:
-            self.print(Text("[Rich]: " + str(e), style="bold red"))
-
-    def _make_plot(
-        self,
-        width,
-        height,
-        epoch: int,
-        train_losses: List[float],
-        val_losses: List[float],
-        model_saver: Optional[BestNModelSaver] = None,
-    ):
-        """Plot the training and validation losses.
-        Args:
-            epoch (int): Current epoch number.
-            train_losses (List[float]): List of training losses.
-            val_losses (List[float]): List of validation losses.
-        Returns:
-            None
-        """
-        if self._plot_log_scale and any(
-            loss_val <= 0 for loss_val in train_losses + val_losses
-        ):
-            raise ValueError(
-                "Cannot plot on a log scale if there are non-positive losses."
-            )
-        plt.clf()
-        plt.plotsize(width, height)
-        plt.title(f"Training curves for {self._run_name}")
-        plt.xlabel("Epoch")
-        plt.theme("dark")
-        if self._plot_log_scale:
-            plt.ylabel("Loss (log scale)")
-            plt.yscale("log")
-        else:
-            plt.ylabel("Loss")
-        plt.grid(True, True)
-
-        plt.plot(
-            list(range(self._starting_epoch, epoch + 1)),
-            train_losses,
-            color=project_conf.Theme.TRAINING.value,
-            # color="blue",
-            label="Training loss",
+        """Return an iterable that tracks the progress of the testing process, and a progress bar
+        hook to update the loss value at each iteration."""
+        return self.query_one(DatasetProgressBar).track_iterable(
+            iterable, Task.TESTING, total
         )
-        plt.plot(
-            list(range(self._starting_epoch, epoch + 1)),
-            val_losses,
-            color=project_conf.Theme.VALIDATION.value,
-            # color="green",
-            label="Validation loss",
-        )
-        best_metrics = (
-            "["
-            + ", ".join(
-                [
-                    f"{metric_name}={metric_value:.2e} "
-                    for metric_name, metric_value in model_saver.best_metrics.items()
-                ]
-                if model_saver is not None
-                else []
-            )
-            + "]"
-        )
-        if model_saver is not None:
-            plt.scatter(
-                [model_saver.min_val_loss_epoch],
-                [model_saver.min_val_loss],
-                color="red",
-                marker="+",
-                label=f"Best model {best_metrics}",
-                style="inverted",
-            )
-        return plt.build()
 
     def plot(
-        self,
-        epoch: int,
-        train_losses: List[float],
-        val_losses: List[float],
-        model_saver: Optional[BestNModelSaver] = None,
+        self, epoch: int, train_loss: float, val_loss: Optional[float] = None
     ) -> None:
-        mk_plot = partial(
-            self._make_plot,
-            epoch=epoch,
-            train_losses=train_losses,
-            val_losses=val_losses,
-            model_saver=model_saver,
-        )
-        self._plot = Panel(PlotextMixin(mk_plot), title="Training curves")
-        self._layout["body"].update(self._plot)
-        self._live.refresh()
+        """Plot the training and validation losses for the current epoch."""
+        self.query_one(PlotterWidget).loading = False
+        self.query_one(PlotterWidget).update(epoch, train_loss, val_loss)
+
+    def set_start_epoch(self, start_epoch: int) -> None:
+        """Set the starting epoch for the plotter widget."""
+        sef = self.query_one(PlotterWidget).set_start_epoch
+
+
+async def run_my_app():
+    gui = GUI("test-run", log_scale=False)
+    task = asyncio.create_task(gui.run_async())
+    while not gui.is_running:
+        await asyncio.sleep(0.01)  # Wait for the app to start up
+    gui.print("Hello, World!")
+    await asyncio.sleep(2)
+    gui.print(Text("Let's log some tensors :)", style="bold magenta"))
+    await asyncio.sleep(0.5)
+    gui.print(torch.rand(2, 4))
+    await asyncio.sleep(2)
+    gui.print(Text("How about some numpy arrays?!", style="italic green"))
+    await asyncio.sleep(1)
+    gui.print(np.random.rand(3, 3))
+    pbar, update_progress_loss = gui.track_training(range(10), 10)
+    for i, e in enumerate(pbar):
+        gui.print(f"[{i+1}/10]: We can iterate over iterables")
+        gui.print(e)
+        await asyncio.sleep(0.1)
+    await asyncio.sleep(2)
+    mnist = MNIST(root="data", train=False, download=True, transform=to_tensor)
+    # Somehow, the dataloader will crash if it's not forked when using multiprocessing along with
+    # Textual.
+    mp.set_start_method("fork")
+    dataloader = DataLoader(mnist, 32, shuffle=True, num_workers=2)
+    pbar, update_progress_loss = gui.track_validation(dataloader, len(dataloader))
+    for i, batch in enumerate(pbar):
+        await asyncio.sleep(0.01)
+        if i % 10 == 0:
+            gui.print(batch)
+            update_progress_loss(random())
+            gui.plot(epoch=i, train_loss=random(), val_loss=random())
+            gui.print(
+                f"[{i+1}/{len(dataloader)}]: We can also iterate over PyTorch dataloaders!"
+            )
+        if i == 0:
+            gui.print(batch)
+    gui.print("Goodbye, world!")
+    _ = await task
 
 
 if __name__ == "__main__":
-    mnist = MNIST(root="data", train=False, download=True, transform=to_tensor)
-    dataloader = DataLoader(mnist, 32, shuffle=True)
-    gui = GUI("test-run", plot_log_scale=False)
-    gui.open()  # TODO: Use a context manager, why not??
-    try:
-        gui.print("Hello, world!")
-        gui.print(
-            "Veeeeeeeeeeeeeeeeeeeryyyyyyyyyyyyyyyy looooooooooooooooooooooooooooong seeeeeeeeeeeenteeeeeeeeeeeeeeennnnnnnnnce!!!!!!!!!!!!!!!!!"
-        )
-        pbar, update_progress_loss = gui.track_training(range(10), 10)
-        for i, e in enumerate(pbar):
-            gui.print(f"[{i}/10]: We can iterate over iterables")
-            sleep(0.1)
-        train_losses, val_losses = [], []
-        pbar, update_progress_loss = gui.track_validation(dataloader, len(dataloader))
-        for i, e in enumerate(pbar):
-            # gui.print(e)  # TODO: Make this work!
-            if i % 10 == 0:
-                train_losses.append(random.random())
-                val_losses.append(random.random())
-                update_progress_loss(random.random())
-                gui.plot(epoch=i, train_losses=train_losses, val_losses=val_losses)
-                gui.print(
-                    f"[{i}/{len(dataloader)}]: We can also iterate over PyTorch dataloaders!"
-                )
-            if i == 0:
-                gui.print(e)
-            sleep(0.01)
-        gui.print("Goodbye, world!")
-        sleep(1)
-    except Exception as e:
-        gui.close()
-        raise e
-    finally:
-        gui.close()
+    asyncio.run(run_my_app())
